requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.3
gunicorn==21.2.0
python-dotenv==1.0.0
numpy>=2.0.0
requests==2.31.0
httpx==0.28.1
anthropic==0.85.0
google-generativeai>=0.8.0
apscheduler==3.11.2
pytz==2025.2
Pillow>=10.0.0
